change syntax in page rank and views and change type in backend.py

+++ src/backend.py
@@ -168,7 +168,7 @@ def search_bm25_body_stem(query: str):
 
 
 # ------- Page Views and Page Rank Functions
-def get_page_view(docs_id: List[str]) -> List[float]:
+def get_page_view(docs_id: List[str]) -> List[int]:
     values = [page_view.get(doc_id, 0) for doc_id in docs_id]
     return values
